X and Y coordinates should be numeric not integers.

+++ R/sql_schema.R
@@ -265,8 +265,8 @@ sql_schema <- list(
         feature_id INTEGER,
         gene_id INTEGER,
         score NUMERIC,
-        x INTEGER,
-        y INTEGER,
+        x NUMERIC,
+        y NUMERIC,
         PRIMARY KEY (id)
       );",
     addSchema = c(
